requirements.txt
langchain
python-dotenv
langchain-community
pypdf
bs4
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu #cpu because of the current device
langchain_chroma
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
pydantic
tf-keras
newspaper3k
arxiv
wikipedia
duckduckgo-search















